refactor(polytope_playground): extract Demo GPU rendering into render module
Moves the render dispatch, point cloud, section-face passes, and the wireframe overlay (~1065 lines) into render.rs as a second impl Demo block. Lifecycle/ui/input methods stay in main.rs.
